test(timeseries): split canvas tests by category, drop redundant exemplars
Extract the shared harness into TimeSeriesPanel.canvasTestUtils.tsx and split the
single canvas test into options/axes/annotations files, each with its own smaller
snapshot (down from one ~10.6k-line snapshot). Remove the Exemplars case, which
only duplicated the defaults snapshot. Same coverage otherwise; also drops the
now-dead clearAxisEvents flag.
